+++ website/src/arcade/useIsDarkMode.ts
@@ -1,0 +1,4 @@
+import * as React from "react";
+import { makeDarkModeHook } from "../../../shared/util/makeDarkModeHook";
+
+export const useIsDarkMode = makeDarkModeHook(React);
